check limits additions
Added function is_in_limits and made sure that checklim function works when called from pyton. is_in_limits takes an array of x,y coordinates and checks if a point is inside the limits provided by xlim, ylim coordinate arrays. This is useful to check which part of a trajectory is inside a closed flux surface